Moved data structures modules under Collections directory and created a Util directory for the paramFrame module.
